// store/modules/user/types.ts
export enum UserType {
  /**
    @types Action
  **/
  ACTION_GET_USERS = 'ACTION_GET_USERS',
  ACTION_GET_USER_BY_AUTH0ID = 'ACTION_GET_USER_BY_AUTH0ID',
  ACTION_GET_USER_DATA_BY_AUTH0ID = 'ACTION_GET_USER_DATA_BY_AUTH0ID',
  ACTION_GET_USER_DEFAULT_CONFIG = 'ACTION_GET_USER_DEFAULT_CONFIG',
  ACTION_UPDATE_USER_PROFILE = 'ACTION_UPDATE_USER_PROFILE',
  ACTION_UPDATE_USER_PERSONAL_INFORMATION = 'ACTION_UPDATE_USER_PERSONAL_INFORMATION',
  ACTION_UPDATE_USER_SETTINGS = 'ACTION_UPDATE_USER_SETTINGS',
  /**
    @types Mutation
  **/
  MUTATION_SET_USERS = 'MUTATION_SET_USERS',
  MUTATION_SET_USER_BY_AUTH0ID = 'MUTATION_SET_USER_BY_AUTH0ID',
  MUTATION_SET_USER_DEFAULT_CONFIG = ' MUTATION_SET_USER_DEFAULT_CONFIG',
  /**
    @types State
  **/
  STATE_LIST_USERS = 'STATE_LIST_USERS',
  STATE_LIST_USER_BY_AUTH0ID = 'STATE_LIST_USER_BY_AUTH0ID',
  STATE_USER_DEFAULT_CONFIG = 'STATE_USER_DEFAULT_CONFIG',
}